tech-news-agent/pyproject.toml
[project]
name = "tech-news-agent"
version = "0.1.0"
description = "データ基盤ニュース配信 LINE bot (RSS / arXiv から収集→LLM 要約→LINE Push)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "feedparser>=6.0.11",
    "arxiv>=2.1.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.2",
    "line-bot-sdk>=3.11.0",
    "llm-client",
    "analytics-platform",
]

[tool.uv.sources]
llm-client = { path = "../llm-client" }
analytics-platform = { path = "../analytics-platform" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.4.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]
